Revert "feat: プレイテストDSL基盤とUI経路E2Eシナリオ・モデル別スキル評価・実バグ修正"

+++ moorestech_server/Assets/Scripts/Game.Block/Blocks/ElectricToGear/ElectricToGearGeneratorComponent.cs
@@ -91,7 +91,7 @@ namespace Game.Block.Blocks.ElectricToGear
 
         #region IElectricConsumer
 
-        public ElectricPower RequestEnergy => new ElectricPower(CurrentMode.RequiredPower);
+        public ElectricPower RequestEnergy => new ElectricPower((float)CurrentMode.RequiredPower);
 
         public void SupplyEnergy(ElectricPower power)
         {
@@ -103,7 +103,7 @@ namespace Game.Block.Blocks.ElectricToGear
         private void UpdateFulfillment(ElectricPower power)
         {
             _suppliedPower = power;
-            var required = CurrentMode.RequiredPower;
+            var required = (float)CurrentMode.RequiredPower;
             // 充足率を計算
             // Calculate fulfillment rate
             var newRate = required > 0f ? Math.Min(power.AsPrimitive() / required, 1f) : 0f;
